Move classes to different package
