+++ game-headed/src/main/java/org/triplea/game/client/ui/javafx/screens/AboutInformation.java
@@ -1,4 +1,4 @@
-package org.triplea.game.client.ui.javafx;
+package org.triplea.game.client.ui.javafx.screens;
 
 import org.triplea.awt.OpenFileUtility;
 import org.triplea.game.client.ui.javafx.screen.ControlledScreen;
