+++ game-headed/src/main/java/org/triplea/game/client/ui/javafx/screens/MainMenuPane.java
@@ -1,4 +1,4 @@
-package org.triplea.game.client.ui.javafx;
+package org.triplea.game.client.ui.javafx.screens;
 
 import java.text.MessageFormat;
