+++ game-headed/src/main/resources/org/triplea/game/client/ui/javafx/fxml/GameSelectionControls.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.layout.HBox?>
 
 
-<BorderPane fx:id="gameOptions" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.triplea.game.client.ui.javafx.GameSelectionControls">
+<BorderPane fx:id="gameOptions" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.triplea.game.client.ui.javafx.screens.GameSelectionControls">
    <center>
       <HBox alignment="CENTER" prefHeight="100.0" prefWidth="200.0">
          <children>
